Move multer config to its own file.
