+++ routes/api/v1/users.js
@@ -1,23 +1,24 @@
 var express           = require('express');
 var router            = express.Router();
 const usersController = require('../../../controllers/api/v1/usersController');
-const multer  = require('multer');
-const storage = multer.diskStorage({
-  destination: (req, file, cb) => {
-    cb(null, 'avatars/')
-  },
-  filename: (req, file, cb) => {
-    cb(null, new Date().toISOString() + file.originalname)
-  }
-});
-const upload  = multer({ storage, limits: {
-  fileSize: 1024 * 1024 * 5
-} });
+// const multer  = require('multer');
+// const storage = multer.diskStorage({
+//   destination: (req, file, cb) => {
+//     cb(null, 'avatars/')
+//   },
+//   filename: (req, file, cb) => {
+//     cb(null, new Date().toISOString() + file.originalname)
+//   }
+// });
+// const upload  = multer({ storage, limits: {
+//   fileSize: 1024 * 1024 * 5
+// } });
+const multer = require('../../../multer');
 
 router.post('/', usersController.create);
 router.get('/', usersController.index);
 router.get('/:uid', usersController.show);
-router.put('/:uid', upload.single('avatar'), usersController.update);
+router.put('/:uid', multer.upload.single('avatar'), usersController.update);
 router.delete('/:uid', usersController.destroy);
 
 module.exports = router;
